requirements.txt
astapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.12
transformers==4.39.3
torch==2.2.2
timm==0.9.12
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.24.4
pytesseract==0.3.10
setuptools>=65.0.0
wheel>=0.41.0
python-multipart==0.0.9
openai>=1.3.0,<2.0.0